Cleanup of diff_eq use + Progress on including integration stages on the GPU
Can now use any differential-equations method included in the gpuable/not_gpuable modules for the integration of any system. Only those that are "gpuable" will be usable with an initialised GPU state.

Get a correct VV implementation working on the GPU and figure out how best to include the required 'stages'.

+++ src/utils.rs
@@ -107,7 +107,7 @@ impl InteractionHandler {
                 WindowEvent::MouseWheel(scroll, _) => {
                     let scale_factor = match scroll {
                         MouseScrollDelta::LineDelta(_, y) => 1.0 + y as f64 * 0.1,
-                        MouseScrollDelta::PixelDelta(pos) => 1.0 + pos.y as f64 * 0.0001,
+                        MouseScrollDelta::PixelDelta(pos) => 1.0 + pos.y * 0.0001,
                     };
                     if app.keys.down.contains(&Key::LShift) {
                         self.dt *= scale_factor;
